requirements.txt
PyQt6
scapy
matplotlib
pandas
pyyaml
jinja2
requests
